requirements.txt
# Smartacus - Amazon Economic Probe
# Python dependencies

# =============================================================================
# Core Dependencies
# =============================================================================

# Keepa API client (official library)
keepa>=1.3.0

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# Environment variable management
python-dotenv>=1.0.0

# =============================================================================
# Data Processing
# =============================================================================

# Data validation and serialization
pydantic>=2.5.0

# Numerical operations
numpy>=1.24.0

# Date/time utilities
python-dateutil>=2.8.2

# =============================================================================
# HTTP and Networking
# =============================================================================

# HTTP client (used by keepa internally)
requests>=2.31.0

# Async HTTP (for future async implementation)
aiohttp>=3.9.0

# =============================================================================
# Logging and Monitoring
# =============================================================================

# Structured logging
structlog>=23.2.0

# Metrics (optional, for production monitoring)
prometheus-client>=0.19.0

# =============================================================================
# Development and Testing
# =============================================================================

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code formatting
black>=23.12.0
isort>=5.13.0

# Linting
flake8>=6.1.0
mypy>=1.8.0

# Type stubs
types-requests>=2.31.0
types-python-dateutil>=2.8.0

# =============================================================================
# Web Framework (API)
# =============================================================================

# FastAPI framework
fastapi>=0.109.0

# ASGI server
uvicorn[standard]>=0.27.0

# =============================================================================
# CLI and Utilities
# =============================================================================

# CLI argument parsing (standard library, but listed for clarity)
# argparse is built-in

# Progress bars for CLI
tqdm>=4.66.0

# Rich console output
rich>=13.7.0

# =============================================================================
# Database Migrations (optional)
# =============================================================================

# Alembic for migrations (if needed beyond raw SQL)
# alembic>=1.13.0
# SQLAlchemy>=2.0.0

# =============================================================================
# AI / LLM
# =============================================================================

# Anthropic Claude SDK
anthropic>=0.40.0

# OpenAI SDK (fallback)
openai>=1.50.0

# =============================================================================
# RAG / Vector Search
# =============================================================================

# pgvector Python client (for vector operations)
# Note: pgvector extension must be installed in PostgreSQL
# CREATE EXTENSION vector;

# =============================================================================
# Scheduling (for production deployment)
# =============================================================================

# Task scheduling
schedule>=1.2.0

# APScheduler for more advanced scheduling
# APScheduler>=3.10.0
